+++ packages/@dicebear/avatars-bottts-sprites/src/sides/cables-01.ts
@@ -1,6 +1,4 @@
-import Color from '@dicebear/avatars/lib/color';
-
-export default (color: Color) => {
+export default (color: string) => {
   return `
         <g opacity="0.9">
             <path id="Cable" d="M38 12C35.046 23.6966 18.0959 18.6663 14.6313 30.009C11.1668 41.3518 22.6565 50 32.1552 50" stroke="#2A3544" stroke-width="6"/>
@@ -11,7 +9,7 @@ export default (color: Color) => {
             <path fill-rule="evenodd" clip-rule="evenodd" d="M138 6C136.895 6 136 6.89543 136 8V22C136 23.1046 136.895 24 138 24H157C158.105 24 159 23.1046 159 22V8C159 6.89543 158.105 6 157 6H138ZM21 37C21 35.8954 21.8954 35 23 35H35C36.1046 35 37 35.8954 37 37V55C37 56.1046 36.1046 57 35 57H23C21.8954 57 21 56.1046 21 55V37ZM136 44C136 42.8954 136.895 42 138 42H157C158.105 42 159 42.8954 159 44V62C159 63.1046 158.105 64 157 64H138C136.895 64 136 63.1046 136 62V44Z" fill="white"/>
         </mask>
         <g mask="url(#sidesCables01Mask0)">
-            <rect width="180" height="76" fill="${color.hex}"/>
+            <rect width="180" height="76" fill="${color}"/>
         </g>
     `;
 };